requirements.txt
openai>=1.0.0
python-dotenv
mcp
httpx>=0.27.2
